+++ angular_test/CHANGELOG.md
@@ -9,6 +9,12 @@
   optionally an `InjectorFactory`. This allows writing tests entirely free
   of any invocations of `initReflector()`.
 
+* **BREAKING CHANGE**: Adding stabilizers to `NgTestBed` now takes a factory
+  function of type `NgTestStabilizer Function(Injector)`, which is aliased as
+  `NgTestStabilizerFactory`. This allows using `NgTestBed` without any dynamic
+  reflective factories (i.e. `initReflector()`) and doesn't have impact to
+  most users.
+
 ## 2.0.0-alpha+3
 
 * Removed support for `pub run angular_test`. This is no longer strictly